Extracting out some repeated functions
