+++ statistics/getGoals.py
@@ -23,14 +23,6 @@ class Player(object):
         self.eflcup = eflcup
         self.total = total
 
-
-def getTimestamp():
-        dt = str(datetime.datetime.now().month) + '/' + str(datetime.datetime.now().day) + ' '
-        hr = str(datetime.datetime.now().hour) if len(str(datetime.datetime.now().hour)) > 1 else '0' + str(datetime.datetime.now().hour)
-        min = str(datetime.datetime.now().minute) if len(str(datetime.datetime.now().minute)) > 1 else '0' + str(datetime.datetime.now().minute)
-        t = '[' + hr + ':' + min + '] '
-        return dt + t
-
 def printPlayers():
     for x in range(len(players)):
         p = players[x]
