+++ table/europaTable.py
@@ -9,13 +9,6 @@ from itertools import groupby
 from time import sleep
 from bs4 import BeautifulSoup
 
-def getTimestamp():
-        dt = str(datetime.datetime.now().month) + '/' + str(datetime.datetime.now().day) + ' '
-        hr = str(datetime.datetime.now().hour) if len(str(datetime.datetime.now().hour)) > 1 else '0' + str(datetime.datetime.now().hour)
-        min = str(datetime.datetime.now().minute) if len(str(datetime.datetime.now().minute)) > 1 else '0' + str(datetime.datetime.now().minute)
-        t = '[' + hr + ':' + min + '] '
-        return dt + t
- 
 def getSprite(teamName):
     return {
          "Arsenal": "(#sprite1-p1)",
